notfication not marking as seen bug fixed
